Refactor : 파일구조 조정
