+++ src/components/detail/index.ts
@@ -1,0 +1,1 @@
+export { default as DetailDiv } from './DetailDiv';
