+++ src/components/project/index.ts
@@ -1,3 +1,2 @@
-export { default as DetailDiv } from './DetailDiv';
 export { default as ProjectCard } from './ProjectCard';
 export { default as ProjectCardDiv } from './ProjectCardDiv';
