+++ src/pages/edit/index.ts
@@ -1,2 +1,2 @@
-export { default as DetailEditPage } from './DetailEditPage';
+export { default as DetailEdit } from './DetailEdit';
 export { default as ProjectEditPage } from './ProjectEditPage';
