+++ src/pages/index.tsx
@@ -4,7 +4,7 @@ import {HomePage, NavPage, PageNotFound} from './components'
 */
 export { HomePage } from './home';
 export { NavPage, ProjectPage, ProjectDetailPage, AboutMePage } from './nav';
-export { DetailEditPage, ProjectEditPage } from './edit';
+export { DetailEdit, ProjectEditPage } from './edit';
 export { PageNotFound } from './PageNotFound';
 export { Login, SignUp, Spinner } from './login';
 export { ContactMe } from './contact/ContactMe';
